feat: improved CAD-settings layout
