+++ apps/client/locales/en/cad-settings.json
@@ -1,20 +1,5 @@
 {
   "CadSettings": {
-    "templateSection": "Templates",
-    "callsignTemplate": "Callsign Template",
-    "callsignTemplateInfo": "This is the template for unit callsigns. You can use the following variables: {variables}",
-    "caseNumberTemplate": "Case Number Template",
-    "caseNumberTemplateInfo": "This is the template for record case numbers. You can use the following variables: {variables}",
-    "pairedUnitTemplate": "Paired Unit Template",
-    "pairedUnitTemplateInfo": "This is the template for paired units. You can use the following variables: {variables}. Note: this will only apply to newly created citizens.",
-    "driversLicenseTemplate": "Drivers license number template",
-    "driversLicenseTemplateInfo": "This is the template for drivers license numbers. You can use the following variables: {variables}. Note: this will only apply to newly created citizens.",
-    "pilotLicenseTemplate": "Pilot license number template",
-    "pilotLicenseTemplateInfo": "This is the template for pilot license numbers. You can use the following variables: {variables}. Note: this will only apply to newly created citizens.",
-    "weaponLicenseTemplate": "Weapon license number template",
-    "weaponLicenseTemplateInfo": "This is the template for weapon license numbers. You can use the following variables: {variables}. Note: this will only apply to newly created citizens.",
-    "waterLicenseTemplate": "Water license number template",
-    "waterLicenseTemplateInfo": "This is the template for water license numbers. You can use the following variables: {variables}. Note: this will only apply to newly created citizens.",
     "generalSettings": "General Settings",
     "cadName": "CAD Name",
     "cadNameDescription": "The name of your SnailyCAD instance. This can be the name of your community or server, etc.",
@@ -129,6 +114,21 @@
     "readMorePublicAPI": "Read more about the <a>Public API here.</a>"
   },
   "MiscSettingsTab": {
+    "templateSection": "Templates",
+    "callsignTemplate": "Callsign Template",
+    "callsignTemplateInfo": "This is the template for unit callsigns. You can use the following variables: {variables}",
+    "caseNumberTemplate": "Case Number Template",
+    "caseNumberTemplateInfo": "This is the template for record case numbers. You can use the following variables: {variables}",
+    "pairedUnitTemplate": "Paired Unit Template",
+    "pairedUnitTemplateInfo": "This is the template for paired units. You can use the following variables: {variables}. Note: this will only apply to newly created citizens.",
+    "driversLicenseTemplate": "Drivers license number template",
+    "driversLicenseTemplateInfo": "This is the template for drivers license numbers. You can use the following variables: {variables}. Note: this will only apply to newly created citizens.",
+    "pilotLicenseTemplate": "Pilot license number template",
+    "pilotLicenseTemplateInfo": "This is the template for pilot license numbers. You can use the following variables: {variables}. Note: this will only apply to newly created citizens.",
+    "weaponLicenseTemplate": "Weapon license number template",
+    "weaponLicenseTemplateInfo": "This is the template for weapon license numbers. You can use the following variables: {variables}. Note: this will only apply to newly created citizens.",
+    "waterLicenseTemplate": "Water license number template",
+    "waterLicenseTemplateInfo": "This is the template for water license numbers. You can use the following variables: {variables}. Note: this will only apply to newly created citizens.",
     "miscSettings": "Miscellaneous Settings",
     "cadRelated": "CAD Related",
     "inactivityTimeouts": "Inactivity Timeouts",
